Restructuring, adding environment file and Sphinx doc.
To edit the docs, edit the docs/build/index.rst then run
make html
You will need Sphinx in your environment; it's available in Conda
