+++ docs/source/_static/tropomi_ut_no2_architecture.png
