+++ uptrop/fresco_cld_err.py
@@ -9,26 +9,23 @@
     (180-450 hPa).
 
     (1) Initial version created on 15 March 2020 by Eloise Marais
-        (eloise.marais@le.ac.uk).
+        (eloise.marais@le.ac.uk). Reworked by John Roberts June 2020
+        (jfr10@le.ac.uk)
 
 '''
 
 # Import relevant packages:
 import glob
-import os
 import numpy as np
-import netCDF4
 from netCDF4 import Dataset
 import matplotlib.pyplot as plt
 from mpl_toolkits.basemap import Basemap
 import argparse
 from os import path
 
-from gamap_colormap import WhGrYlRd
+from uptrop.gamap_colormap import WhGrYlRd
 
-import pdb
-
-from convert_height_to_press import alt2pres
+from uptrop.convert_height_to_press import alt2pres
 
 # Turn off warnings:
 # np.warnings.filterwarnings('ignore')
